backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
requests>=2.31
python-dotenv>=1.0
